fix(Tracking): account for tracked object rotation in angular velocity
The angular velocities reported by the Oculus Plugin don't seem to
account for any rotations of the object being tracked and therefore
the returned angular velocity is incorrect.

The fix is to just use the current rotation of the object to track as
an offset of the current reported angular velocity.