fix(Tracking): account for camera rig direction when reporting velocity
The rotation of the CameraRig direction will cause the velocities to be
reported incorrectly as they are not taking into account the offset of
the new direction being faced.

The fix is to provide a new relativeTo field that allows specifying the
play area (OVRCameraRig) that will utilize the rotation of that
GameObject in offsetting the velocity calculation.